requirements.txt
discord.py>=1.6.0
inspiro>=0.0.2
pillow>=8.1.0
psutil>=5.8.0
PyNaCl>=1.4.0
requests>=2.25.1
bs4>=0.0.1
mpmath>=1.1.0
numpy>=1.19.4
PyNaCl>=1.4.0
python-dateutil>=2.8.1
python-magic-bin>=0.4.14
pytz>=2020.4
requests>=2.24.0
sympy>=1.6.2
youtube-dlc>=2020.11.11.post3
asyncio>=3.4.3